refactor: Fix typo
