+++ frontend/src/components/BlockEditor.vue
@@ -99,7 +99,7 @@ const rotating = ref(false);
 const transforming = computed(() => resizing.value || rotating.value);
 const guides = setGuides(props.target, canvasProps);
 const moving = ref(false);
-const preventCLick = ref(false);
+const preventClick = ref(false);
 
 const showPaddingHandler = computed(() => {
 	return (
@@ -230,8 +230,8 @@ onMounted(() => {
 
 const handleClick = (ev: MouseEvent) => {
 	if (props.editable || props.readonly) return;
-	if (preventCLick.value) {
-		preventCLick.value = false;
+	if (preventClick.value) {
+		preventClick.value = false;
 		return;
 	}
 
@@ -326,7 +326,7 @@ const handleMove = (ev: MouseEvent) => {
 		}
 
 		mouseMoveEvent.preventDefault();
-		preventCLick.value = true;
+		preventClick.value = true;
 	};
 	document.addEventListener("mousemove", mousemove);
 	document.addEventListener(
