+++ frontend/src/components/BlockPropertySections/DimensionSection.ts
